More CGFloat, NSUInteger and NSInteger changes.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36154 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSComboBox.h
@@ -45,8 +45,8 @@
 - (NSSize) intercellSpacing;
 - (void) setIntercellSpacing: (NSSize)aSize;
 
-- (float) itemHeight;
-- (void) setItemHeight: (float)itemHeight;
+- (CGFloat) itemHeight;
+- (void) setItemHeight: (CGFloat)itemHeight;
 
 - (NSInteger) numberOfVisibleItems;
 - (void) setNumberOfVisibleItems: (NSInteger)visibleItems;
